PROJECT-breastfeeding_analysis/requirements.txt
google-cloud-storage
google-cloud-bigquery
prefect
requests
dbt-bigquery
dbt-core
